@@ -1,39 +1,35 @@
 {
-  "name": "binahria-webpage",
-  "version": "0.1.1",
+  "name": "binaria-final-page",
+  "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@fortawesome/fontawesome-svg-core": "^6.1.1",
-    "@fortawesome/free-brands-svg-icons": "^6.1.1",
-    "@fortawesome/free-solid-svg-icons": "^6.1.1",
-    "@fortawesome/react-fontawesome": "^0.1.18",
-    "@testing-library/jest-dom": "^5.13.0",
-    "@testing-library/react": "^11.2.7",
-    "@testing-library/user-event": "^12.8.3",
-    "@typeform/embed-react": "^1.13.0",
-    "axios": "^1.1.0",
-    "firebase": "^9.6.9",
+    "@fortawesome/fontawesome-svg-core": "^6.2.1",
+    "@fortawesome/free-brands-svg-icons": "^6.2.1",
+    "@fortawesome/free-solid-svg-icons": "^6.2.1",
+    "@fortawesome/react-fontawesome": "^0.2.0",
+    "@testing-library/jest-dom": "^5.16.5",
+    "@testing-library/react": "^13.4.0",
+    "@testing-library/user-event": "^13.5.0",
+    "@typeform/embed-react": "^2.4.0",
+    "axios": "^1.2.0",
     "gsap": "^3.11.3",
-    "i18next": "^21.8.4",
-    "i18next-browser-languagedetector": "^6.1.4",
-    "react": "^17.0.2",
-    "react-animated-cursor": "^2.2.0",
+    "i18next": "^22.0.6",
+    "i18next-browser-languagedetector": "^7.0.1",
+    "react": "^18.2.0",
     "react-calendly": "^4.1.0",
-    "react-dom": "^17.0.2",
+    "react-dom": "^18.2.0",
     "react-hot-toast": "^2.4.0",
-    "react-i18next": "^11.16.9",
-    "react-multi-carousel": "^2.6.3",
-    "react-player": "^2.9.0",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "4.0.3",
+    "react-i18next": "^12.0.0",
+    "react-multi-carousel": "^2.8.2",
+    "react-player": "^2.11.0",
+    "react-router-dom": "^6.4.3",
+    "react-scripts": "5.0.1",
     "react-tabs": "^5.1.0",
-    "react-tsparticles": "^1.27.0",
-    "react-vis": "^1.11.7",
+    "react-tsparticles": "^2.5.3",
     "reactjs-popup": "^2.0.5",
-    "swiper": "^8.3.2",
-    "web-vitals": "^1.1.2"
+    "swiper": "^8.4.5",
+    "web-vitals": "^2.1.4"
   },
-  "homepage": "https://binahria.com.mx/",
   "scripts": {
     "start": "react-scripts start",
     "build": "react-scripts build",
@@ -57,13 +53,5 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
-  },
-  "description": "TODO: Give a short introduction of your project. Let this section explain the objectives or the motivation behind this project.",
-  "main": "index.js",
-  "repository": {
-    "type": "git",
-    "url": "https://binahriaanalytics@dev.azure.com/binahriaanalytics/P%C3%A1gina%20web/_git/Pagina_web_2"
-  },
-  "author": "",
-  "license": "ISC"
+  }
 }
